# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NodeStatus = (typeof NodeStatus)[keyof typeof NodeStatus];

export const NodeStatus = {
  healthy: "healthy",
  degraded: "degraded",
  failing: "failing",
  repairing: "repairing",
  offline: "offline",
} as const;

export interface GridNode {
  id: string;
  name: string;
  status: NodeStatus;
  /** Latency in ms */
  latency: number;
  /** Error rate 0-1 */
  errorRate: number;
  /** Uptime percentage 0-100 */
  uptime: number;
  /** CPU utilization 0-100% */
  cpu: number;
  /** Memory utilization 0-100% */
  memory: number;
  /** Network ingress in Mbps */
  networkIn: number;
  /** Network egress in Mbps */
  networkOut: number;
  lastUpdated: string;
  assignedAgent?: string | null;
}

export interface GridState {
  nodes: GridNode[];
  timestamp: string;
}

export type InjectShockBodySeverity =
  (typeof InjectShockBodySeverity)[keyof typeof InjectShockBodySeverity];

export const InjectShockBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  catastrophic: "catastrophic",
} as const;

export interface InjectShockBody {
  severity?: InjectShockBodySeverity;
  targetNodeIds?: string[];
}

export interface ShockResult {
  affectedNodes: string[];
  severity: string;
  message: string;
}

export interface NodeRepairResult {
  nodeId: string;
  status: string;
  message: string;
}

export interface GridSummary {
  totalNodes: number;
  healthyNodes: number;
  degradedNodes: number;
  failingNodes: number;
  repairingNodes: number;
  offlineNodes: number;
  overallHealthPercent: number;
  avgLatency: number;
  activeIncidents: number;
}

export type AgentLogEntryAgent =
  (typeof AgentLogEntryAgent)[keyof typeof AgentLogEntryAgent];

export const AgentLogEntryAgent = {
  sentinel: "sentinel",
  coordinator: "coordinator",
  diagnostician: "diagnostician",
  remediator: "remediator",
  validator: "validator",
  engineer: "engineer",
  system: "system",
} as const;

export type AgentLogEntryLevel =
  (typeof AgentLogEntryLevel)[keyof typeof AgentLogEntryLevel];

export const AgentLogEntryLevel = {
  info: "info",
  warning: "warning",
  critical: "critical",
  action: "action",
  success: "success",
  report: "report",
} as const;

export type AgentLogEntryMetadata = { [key: string]: unknown } | null;

export interface AgentLogEntry {
  id: string;
  timestamp: string;
  agent: AgentLogEntryAgent;
  level: AgentLogEntryLevel;
  message: string;
  nodeId?: string | null;
  metadata?: AgentLogEntryMetadata;
}

export interface AgentLogsResponse {
  logs: AgentLogEntry[];
  total: number;
}

export interface ScanResult {
  nodesScanned: number;
  issuesFound: number;
  actionsTriggered: number;
  summary: string;
}

export type ApprovalRequestRiskLevel =
  (typeof ApprovalRequestRiskLevel)[keyof typeof ApprovalRequestRiskLevel];

export const ApprovalRequestRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ApprovalRequest {
  id: string;
  nodeId: string;
  nodeName: string;
  action: string;
  infraCommands: string[];
  justification: string;
  confidence: number;
  riskLevel: ApprovalRequestRiskLevel;
  requestedBy: string;
  timestamp: string;
}

export interface ApprovalsResponse {
  approvals: ApprovalRequest[];
}

export interface ApprovalDecision {
  approved: boolean;
  id: string;
}

export type GetAgentLogsParams = {
  limit?: number;
};