# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/approvals/{id}/approve:
    post:
      operationId: approveAction
      tags: [agents]
      summary: Approve a pending remediation action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecision"
  /agents/approvals/{id}/reject:
    post:
      operationId: rejectAction
      tags: [agents]
      summary: Reject a pending remediation action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecision"
components:
  schemas:
...
    ApprovalDecision:
      type: object
      properties:
        approved:
          type: boolean
        id:
          type: string
      required:
        - approved
        - id